v14: Make state field optional and disallow null for state.open
[breaking-change]:
In v0.13 state and but state.open must be present, but state.open may be
null. This doesn't provide any benefit, so v14 removed these
restrictions.

We can't provide full backwards compatiblity, since state.open == null
is not supported in v14.

To still have a json which is compatible with both v0.13 and v14
state.open must be set to either true or false.